move the drop-logging infra to a separate crate so it is easier to use everywhere and replace eprintln! by error!
