+++ raphtory/src/arrow_loader/df_loaders/edges.rs
@@ -57,6 +57,7 @@ use storage::{
     },
     Extension,
 };
+use tracing::error;
 
 #[derive(Debug, Copy, Clone)]
 pub struct ColumnNames<'a> {
@@ -130,7 +131,7 @@ pub fn load_edges_from_df_prefetch<
                 let sender = tx;
                 for chunk in chunks {
                     if let Err(e) = sender.send(chunk) {
-                        eprintln!("Error sending chunk to loader: {}", e);
+                        error!("Error sending chunk to loader: {}", e);
                         break;
                     }
                 }
